+++ src/main/java/lotto/domain/lotto/PurchasePrice.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import lotto.exception.ErrorMessagePrinter;
 import lotto.exception.LottoExceptionMessage;
