+++ src/main/java/lotto/domain/lotto/generator/AutoGenerator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.domain.lotto.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
