+++ src/main/java/lotto/domain/lottoResult/LottoResultStatus.java
@@ -1,4 +1,4 @@
-package lotto.constant;
+package lotto.domain.lottoResult;
 
 import java.util.stream.Stream;
