+++ src/main/java/lotto/domain/lottoWinning/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottoWinning;
 
 import java.util.HashSet;
 import java.util.List;
