+++ src/test/java/lotto/constant/LottoResultStatusTest.java
@@ -2,6 +2,7 @@ package lotto.constant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lotto.domain.lottoResult.LottoResultStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
