+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -3,8 +3,14 @@ package lotto.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import lotto.constant.LottoResultStatus;
-import lotto.util.ManualGenerator;
+import lotto.domain.lotto.PurchaseLottos;
+import lotto.domain.lotto.PurchasePrice;
+import lotto.domain.lotto.generator.ManualGenerator;
+import lotto.domain.lottoResult.LottoResult;
+import lotto.domain.lottoResult.LottoResultStatus;
+import lotto.domain.lottoWinning.BonusNumber;
+import lotto.domain.lottoWinning.LottoWinning;
+import lotto.domain.lottoWinning.WinningNumbers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
