+++ src/test/java/lotto/domain/LottoTest.java
@@ -4,7 +4,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import lotto.constant.LottoResultStatus;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lottoResult.LottoResultStatus;
+import lotto.domain.lottoWinning.BonusNumber;
+import lotto.domain.lottoWinning.LottoWinning;
+import lotto.domain.lottoWinning.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
