+++ src/test/java/lotto/domain/LottosTest.java
@@ -2,7 +2,8 @@ package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.util.AutoGenerator;
+import lotto.domain.lotto.Lottos;
+import lotto.domain.lotto.generator.AutoGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
